🚀 2단계 - 깃헙 로그인 구현 (#617)
* fix : PathFinder 수정

* feature: 즐겨찾기 밸리데이션 추가

* feature: 2단계 구현 완료

+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -1,8 +1,9 @@
 package nextstep.favorite.application;
 
+import nextstep.favorite.application.exceptions.BadRequestException;
 import nextstep.favorite.application.exceptions.CannotFavoriteStationException;
 import nextstep.favorite.application.exceptions.FavoriteNotFoundException;
-import nextstep.member.MemberNotFoundException;
+import nextstep.member.exceptions.MemberNotFoundException;
 import nextstep.favorite.application.dto.FavoriteRequest;
 import nextstep.favorite.application.dto.FavoriteResponse;
 import nextstep.favorite.dao.StationDao;
@@ -39,30 +40,22 @@ public class FavoriteService {
         this.pathService = pathService;
     }
 
-    /**
-     * TODO: LoginMember 를 추가로 받아서 FavoriteRequest 내용과 함께 Favorite 를 생성합니다.
-     *
-     * @param request
-     */
+
     public FavoriteResponse createFavorite(LoginMember loginMember, FavoriteRequest request) {
         Member member = getMember(loginMember.getEmail());
 
         Station source = stationDao.findStation(request.getSource());
         Station target = stationDao.findStation(request.getTarget());
+
         if (!pathService.isConnectedPath(source, target)) {
             throw new CannotFavoriteStationException("연결된 역이 아닙니다");
         }
 
-        Favorite favorite = new Favorite(source, target, member);
-        Favorite saved = favoriteRepository.save(favorite);
+        Favorite saved = favoriteRepository.save(new Favorite(source, target, member));
         return new FavoriteResponse(saved.getId());
     }
 
-    /**
-     * TODO: StationResponse 를 응답하는 FavoriteResponse 로 변환해야 합니다.
-     *
-     * @return
-     */
+
     public List<FavoriteResponse> findFavorites(LoginMember loginMember) {
         Member member = getMember(loginMember.getEmail());
         List<Favorite> favorites = favoriteRepository.findByMember(member);
@@ -75,15 +68,16 @@ public class FavoriteService {
                 .collect(Collectors.toList());
     }
 
-    /**
-     * TODO: 요구사항 설명에 맞게 수정합니다.
-     *
-     * @param loginMember
-     * @param id
-     */
+
     public void deleteFavorite(LoginMember loginMember, Long id) {
         Member member = getMember(loginMember.getEmail());
-        favoriteRepository.deleteByMemberAndId(member, id);
+        favoriteRepository.findById(id)
+            .ifPresent(favorite -> {
+                if (!favorite.isOwner(member)) {
+                    throw new BadRequestException("다른 회원의 즐겨찾기를 삭제할 수 없어요.");
+                }
+                favoriteRepository.deleteByMemberAndId(member, id);
+        });
     }
 
     public FavoriteResponse findFavorite(long favoriteId) {
@@ -96,6 +90,7 @@ public class FavoriteService {
                 new StationResponse(favorite.getTarget().getId(), favorite.getTarget().getName())
             );
     }
+
 
     @Transactional(readOnly = true)
     public Member getMember(String email) {
